add apiRouter, move it to routes/index
