+++ server/package.json
@@ -22,7 +22,7 @@
     "export": "node scripts/export.js",
     "watch": "NODE_ENV=local nodemon main.js",
     "start": "NODE_ENV=production node main.js",
-    "test": "NODE_ENV=test jest --runInBand --forceExit --detectOpenHandles"
+    "test": "NODE_ENV=test jest --silent --runInBand --forceExit --detectOpenHandles"
   },
   "dependencies": {
     "axios": "0.21.1",
